% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_choices}
\alias{select_choices}
\title{Create Select Choices}
\usage{
select_choices(choices)
}
\arguments{
\item{choices}{Choices, named vector or list.}
}
\description{
Create choices for a select input.
}
\keyword{internal}
