% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server}
\usage{
server(input, output, session)
}
\arguments{
\item{input, output}{Environments.}

\item{session}{Shiny session.}
}
\description{
Server logic.
}
\keyword{internal}
